feat: clean up unused util exports
